refactor: change EmailInput to SignIn section
