+++ src/pages/donations/CausesPage/NonProfitsList/index.tsx
@@ -71,6 +71,7 @@ function NonProfitsList({
       newLogEvent("click", "P1_donateBtn", {
         nonProfitId: nonProfit.id,
       });
+      navigateTo({ pathname: "confirm-donation", state: { nonProfit } });
       setConfirmModalVisible(true);
     } else {
       newLogEvent("click", "P1_donateBlockedBtn", {
