+++ src/pages/donations/ConfirmDonationPage/index.tsx
@@ -8,7 +8,7 @@ import RightImage from "assets/images/top-right-sun-shape.svg";
 import DonatingSection from "pages/donations/ConfirmDonationPage/DonatingSection";
 import useDonationFlow from "hooks/useDonationFlow";
 import useNavigation from "hooks/useNavigation";
-import EmailInputSection from "pages/donations/ConfirmDonationPage/EmailInputSection";
+import SignInSection from "pages/donations/ConfirmDonationPage/SignInSection";
 import * as S from "./styles";
 
 type LocationStateType = {
@@ -61,7 +61,7 @@ function ConfirmDonationPage(): JSX.Element {
       ) : signedIn ? (
         <SignedInSection nonProfit={nonProfit} onContinue={onContinue} />
       ) : (
-        <EmailInputSection nonProfit={nonProfit} onContinue={onContinue} />
+        <SignInSection nonProfit={nonProfit} onContinue={onContinue} />
       )}
     </S.Container>
   );
